chore: cleanup windows installer
