+++ src/ansys/tools/installer/installer.py
@@ -22,41 +22,8 @@
 
 """Installer module for Ansys Python Manager."""
 
-import logging
-import subprocess
-
 from ansys.tools.installer.linux_functions import install_python_linux, is_linux_os
-
-LOG = logging.getLogger(__name__)
-
-
-def run_ps(command, full_path_to_ps=False):
-    """Run a powershell command as admin."""
-    ps = (
-        r"C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe"
-        if full_path_to_ps
-        else "powershell.exe"
-    )
-    ps_command = [ps, command]
-
-    LOG.debug("Running: %s", str(ps_command))
-    proc = subprocess.run(
-        ps_command,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-        shell=True,
-    )
-
-    # stderr goes through stdout too
-    if proc.returncode:  # If return code != 0
-        LOG.error("From powershell: %s", proc.stdout)
-        if full_path_to_ps == False:
-            return run_ps(command, full_path_to_ps=True)
-
-    else:
-        LOG.debug("From powershell: %s", proc.stdout)
-
-    return proc.stdout, proc.returncode
+from ansys.tools.installer.windows_functions import install_python_windows
 
 
 def install_python(filename, wait=True):
@@ -65,6 +32,4 @@ def install_python(filename, wait=True):
         install_python_linux(filename)
         return "Success", None
     else:
-        wait_str = " -Wait" if wait else ""
-        command = f"(Start-Process '{filename}' -ArgumentList '/passive InstallAllUsers=0' {wait_str})"
-        return run_ps(command)
+        return install_python_windows(filename, wait)
